position = delta * velocity

+++ gameserver/src/engine/components/physics.rs
@@ -3,13 +3,13 @@ use specs::{Component, VecStorage};
 #[derive(Component, Debug)]
 #[storage(VecStorage)]
 pub struct Position {
-    pub x: f32,
-    pub y: f32
+    pub x: f64,
+    pub y: f64
 }
 
 #[derive(Component, Debug)]
 #[storage(VecStorage)]
 pub struct Velocity {
-    pub x: f32,
-    pub y: f32
+    pub x: f64,
+    pub y: f64
 }
